Move CursorProcessor to project package
