+++ presto-main/src/main/java/com/facebook/presto/operator/project/InterpretedCursorProcessor.java
@@ -15,7 +15,6 @@ package com.facebook.presto.operator.project;
 
 import com.facebook.presto.Session;
 import com.facebook.presto.metadata.Metadata;
-import com.facebook.presto.operator.CursorProcessor;
 import com.facebook.presto.spi.ConnectorSession;
 import com.facebook.presto.spi.PageBuilder;
 import com.facebook.presto.spi.RecordCursor;
